fixed behaviorrefueling

+++ Start-a-Town!/Start-a-Town!/Components/Interactions/NewScripts/InteractionToolUse.cs
@@ -91,7 +91,7 @@ namespace Start_a_Town_
             var strength = this.Actor[AttributeDefOf.Strength].Level;
             var fromToolWeight = //10 * 
                 toolWeight / strength;
-            return (int)fromToolWeight;
+            return fromToolWeight;
         }
 
         protected abstract float Progress { get; }
